Revert "Merge pull request #890 from YasinHossain/perf/remove-framer-ssr-tafsir"
This reverts commit 61212237c8e1a509938acf1bf35604d0d8bef541, reversing
changes made to 02cbf71bddc43419698fb6b3b8da7a0b917088c8.
