+++ app/(features)/tafsir/hooks/useTafsirVerseData.ts
@@ -2,7 +2,6 @@ import { useCallback, useEffect, useMemo } from 'react';
 import useSWR, { useSWRConfig } from 'swr';
 
 import { usePrefetchSingleVerse, useSingleVerse } from '@/app/shared/hooks/useSingleVerse';
-import { useSettings } from '@/app/providers/SettingsContext';
 import { GetTafsirContentUseCase } from '@/src/application/use-cases/GetTafsirContent';
 import { container } from '@/src/infrastructure/di/Container';
 import { Surah, Verse as VerseType } from '@/types';
@@ -31,17 +30,8 @@ interface UseTafsirVerseDataReturn {
 
 const PREFETCH_IDLE_DELAY_MS = 400;
 
-export const useTafsirVerseData = (
-  surahId: string,
-  ayahId: string,
-  initial?: {
-    verse?: VerseType | undefined;
-    tafsirHtml?: string | undefined;
-    tafsirResource?: { id: number; name: string; lang: string } | undefined;
-  }
-): UseTafsirVerseDataReturn => {
+export const useTafsirVerseData = (surahId: string, ayahId: string): UseTafsirVerseDataReturn => {
   const { mutate } = useSWRConfig();
-  const { settings } = useSettings();
   const { translationOptions, selectedTranslationName } = useTranslationOptions();
   const { tafsirResource, selectedTafsirName } = useTafsirOptions();
   const { wordLanguageOptions, wordLanguageMap, selectedWordLanguageName, resetWordSettings } =
@@ -49,24 +39,12 @@ export const useTafsirVerseData = (
   const { prev, next, navigate, currentSurah } = useVerseNavigation(surahId, ayahId);
 
   const verseKey = surahId && ayahId ? `${surahId}:${ayahId}` : '';
-  const { verse } = useSingleVerse({
-    idOrKey: verseKey,
-    ...(initial?.verse ? { initialVerse: initial.verse } : {}),
-  });
+  const { verse } = useSingleVerse({ idOrKey: verseKey });
 
-  const repository = useMemo(() => container.getTafsirRepository(), []);
+  const repository = container.getTafsirRepository();
   const tafsirUseCase = useMemo(() => new GetTafsirContentUseCase(repository), [repository]);
 
   const prefetchSingleVerse = usePrefetchSingleVerse();
-
-  const activeTafsirId = settings.tafsirIds?.[0];
-  const initialTafsirId = initial?.tafsirResource?.id;
-  const initialVerseKey = initial?.verse?.verse_key;
-  const effectiveTafsirResource =
-    tafsirResource ??
-    (typeof activeTafsirId === 'number' && initialTafsirId === activeTafsirId
-      ? initial?.tafsirResource
-      : undefined);
 
   const neighborKeys = useMemo(() => {
     const keys: string[] = [];
@@ -81,68 +59,40 @@ export const useTafsirVerseData = (
 
   const prefetchTafsir = useCallback(
     async (keys: string[]) => {
-      if (!activeTafsirId) return;
+      if (!tafsirResource) return;
       await Promise.all(
         keys.map((key) =>
           mutate(
-            ['tafsir', key, activeTafsirId],
-            () => tafsirUseCase.execute(key, activeTafsirId),
+            ['tafsir', key, tafsirResource.id],
+            () => tafsirUseCase.execute(key, tafsirResource.id),
             { populateCache: true, revalidate: false }
           ).catch(() => {})
         )
       );
     },
-    [activeTafsirId, mutate, tafsirUseCase]
+    [mutate, tafsirResource, tafsirUseCase]
   );
 
   const { data: tafsirHtml } = useSWR(
-    verse && activeTafsirId ? ['tafsir', verse.verse_key, activeTafsirId] : null,
-    ([, key, id]) => tafsirUseCase.execute(key as string, id as number),
-    {
-      ...(initial?.tafsirHtml &&
-      initialTafsirId === activeTafsirId &&
-      typeof initialVerseKey === 'string' &&
-      verseKey === initialVerseKey
-        ? { fallbackData: initial.tafsirHtml }
-        : {}),
-    }
+    verse && tafsirResource ? ['tafsir', verse.verse_key, tafsirResource.id] : null,
+    ([, key, id]) => tafsirUseCase.execute(key as string, id as number)
   );
 
   useEffect(() => {
     if (!verse?.verse_key || neighborKeys.length === 0) return;
 
-    const connection = (navigator as unknown as { connection?: { saveData?: boolean } }).connection;
-    if (connection?.saveData) return;
-
-    const schedule = (work: () => void) => {
-      const idleCallback = (window as unknown as { requestIdleCallback?: unknown }).requestIdleCallback;
-      if (typeof idleCallback === 'function') {
-        const handle = (idleCallback as (cb: () => void, options?: { timeout: number }) => number)(
-          work,
-          { timeout: PREFETCH_IDLE_DELAY_MS * 4 }
-        );
-        const cancelIdle = (window as unknown as { cancelIdleCallback?: unknown }).cancelIdleCallback;
-        if (typeof cancelIdle === 'function') {
-          return () => (cancelIdle as (id: number) => void)(handle);
-        }
-        return () => {};
-      }
-      const timer = window.setTimeout(work, PREFETCH_IDLE_DELAY_MS);
-      return () => window.clearTimeout(timer);
-    };
-
-    const cancel = schedule(() => {
+    const timer = window.setTimeout(() => {
       void prefetchSingleVerse(neighborKeys);
       void prefetchTafsir(neighborKeys);
-    });
+    }, PREFETCH_IDLE_DELAY_MS);
 
-    return cancel;
+    return () => window.clearTimeout(timer);
   }, [neighborKeys, prefetchSingleVerse, prefetchTafsir, verse?.verse_key]);
 
   return {
     verse,
     tafsirHtml,
-    tafsirResource: effectiveTafsirResource,
+    tafsirResource,
     translationOptions,
     wordLanguageOptions,
     wordLanguageMap,
