+++ app/shared/components/modal/__tests__/UnifiedModal.test.tsx
@@ -3,6 +3,23 @@ import React from 'react';
 
 import { UnifiedModal } from '@/app/shared/components/modal/UnifiedModal';
 
+// Mock framer-motion to avoid animation timing issues in tests
+jest.mock('framer-motion', () => ({
+  motion: {
+    div: ({ children, ...props }: any) => {
+      // Strip motion-only props to avoid React warnings
+      const domProps = { ...props };
+      delete domProps.variants;
+      delete domProps.initial;
+      delete domProps.animate;
+      delete domProps.exit;
+      delete domProps.transition;
+      return React.createElement('div', domProps, children);
+    },
+  },
+  AnimatePresence: ({ children }: any) => <>{children}</>,
+}));
+
 describe('UnifiedModal', () => {
   it('locks page scrolling while open', async () => {
     const previousHtmlOverflow = document.documentElement.style.overflow;
