+++ app/shared/icons/index.ts
@@ -5,7 +5,6 @@ export { ArrowLeftIcon } from './ArrowLeftIcon';
 export { ChevronDownIcon } from './ChevronDownIcon';
 export { ChevronLeftIcon } from './ChevronLeftIcon';
 export { ChevronRightIcon } from './ChevronRightIcon';
-export { ChevronUpIcon } from './ChevronUpIcon';
 export { GoToIcon } from './GoToIcon';
 
 // Media Controls
