+++ app/shared/player/components/ReciterPanel.tsx
@@ -1,3 +1,4 @@
+import { motion } from 'framer-motion';
 import React, { memo } from 'react';
 import { useTranslation } from 'react-i18next';
 
@@ -31,7 +32,7 @@ export const ReciterPanel = memo(function ReciterPanel({
           {reciters.map((r) => {
             const isSelected = localReciter === r.id;
             return (
-              <button
+              <motion.button
                 key={r.id}
                 onClick={() => setLocalReciter(r.id)}
                 className={cn(
@@ -42,6 +43,7 @@ export const ReciterPanel = memo(function ReciterPanel({
                   touchClasses.target,
                   touchClasses.focus
                 )}
+                whileTap={{ scale: 0.98 }}
               >
                 <div className="min-w-0 flex-1">
                   <div className="text-sm font-medium truncate text-foreground">{r.name}</div>
@@ -58,7 +60,7 @@ export const ReciterPanel = memo(function ReciterPanel({
                     }`}
                   />
                 </div>
-              </button>
+              </motion.button>
             );
           })}
         </div>
