+++ app/shared/ui/cards/FolderProgress.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import { motion } from 'framer-motion';
 import React, { memo } from 'react';
 
 interface FolderProgressProps {
@@ -9,13 +10,14 @@ interface FolderProgressProps {
 export const FolderProgress = memo(function FolderProgress({
   bookmarkCount,
 }: FolderProgressProps): React.JSX.Element {
-  const width = bookmarkCount > 0 ? '100%' : '25%';
   return (
     <div>
       <div className="w-full h-1.5 bg-surface-hover rounded-full overflow-hidden">
-        <div
+        <motion.div
+          initial={{ width: 0 }}
+          animate={{ width: bookmarkCount > 0 ? '100%' : '25%' }}
+          transition={{ delay: 0.2, duration: 0.6, ease: 'easeOut' }}
           className="h-full bg-gradient-to-r from-accent to-accent/80 rounded-full"
-          style={{ width }}
         />
       </div>
       <div className="mt-3 flex items-center justify-between text-xs text-muted">
