+++ app/shared/verse-planner-modal/PlannerSelectionCard.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import { motion } from 'framer-motion';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 
@@ -52,16 +53,17 @@ export function PlannerSelectionCard({
     : 'text-sm leading-snug text-muted/80';
 
   return (
-    <button
+    <motion.button
       type="button"
       onClick={onSelect}
       className={composedClassName}
       aria-pressed={isSelected}
+      whileTap={{ scale: 0.98 }}
     >
       <div className="flex flex-col gap-1">
         <span className={titleClassName}>{localizedPlanName}</span>
         <p className={detailClassName}>{detailLine}</p>
       </div>
-    </button>
+    </motion.button>
   );
 }
