+++ package.json
@@ -85,6 +85,7 @@
     "@vercel/analytics": "^1.6.1",
     "@vercel/speed-insights": "^1.3.1",
     "clsx": "^2.1.1",
+    "framer-motion": "^12.23.12",
     "i18next": "^25.6.2",
     "isomorphic-dompurify": "^2.26.0",
     "lucide-react": "^0.548.0",
